requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
chromadb
scikit-learn
tqdm
PyPDF2
pycryptodome
